+++ lib/rubocop/cop/rspec/be_eql.rb
@@ -39,6 +39,7 @@ module RuboCop
         extend AutoCorrector
 
         MSG = 'Prefer `be` over `eql`.'
+        RESTRICT_ON_SEND = %i[to].freeze
 
         def_node_matcher :eql_type_with_identity, <<-PATTERN
           (send _ :to $(send nil? :eql {true false int float sym nil_type?}))